security/nss: unbreak on powerpc64 after r405494
mpi/mpcpucache.c:728:23: error: incompatible pointer types passing 'int *' to parameter of type 'size_t *' (aka 'unsigned long *') [-Werror,-Wincompatible-pointer-types]
            &cacheline_size, &clen, NULL, 0) < 0 || !cacheline_size)
                             ^~~~~
/usr/include/sys/sysctl.h:1062:48: note: passing argument to parameter here
int     sysctl(const int *, u_int, void *, size_t *, const void *, size_t);
                                                   ^
PR:		147696 (typo origin)
Reported by:	Mark Millard

+++ security/nss/files/patch-lib_freebl_mpi_mpcpucache.c
@@ -17,7 +17,7 @@
 +{
 +	static int cacheline_size = 0;
 +	static int cachemib[] = { CTL_MACHDEP, CPU_CACHELINE };
-+	int clen;
++	size_t clen;
 +	
 +	if (cacheline_size > 0)
 +		return cacheline_size;
